Refactor annotation shortcut handling
* Using X.Y.Z.9000 version to indicate development versions.
* New functions in utils to deal with annotations.
* check_annotations() looks for unsupported annotations against
supported_annotations(), existence of basicgenes and detailedgenes
shortcuts, the same genome prefix in all annotations.
* expand_annotations() takes a list of annotations, looks for any
shortcuts, and expands them with the proper annotations, in the
upstream to downstream order. Shortcuts references are then removed and
the new annotation vector is returned. This function is only designed
to be run after check_annotations() to ensure there are no underlying
problems.
* Shortcut accessors have been added to supported_annotations() with
the caveat that shortcut accessors cannot be used in data(...).
* Shortcut accessors should now have the genome appended to the front,
e.g. hg19_cpgs, mm10_detailedgenes, or mm9_basicgenes. This fits with
the general naming style of the annotations.
* genome parameter removed from annotate_regions() because of the new
shortcut naming scheme.
* Documentation, examples, tests, and vignette have been updated to
reflect these changes.
* Removed tests from functions that use check_annotations().
* Updated README.md to reflect vignette changes.