Moving version back to v0.2.0
I'm discovering that the API and features for annotatr are not nearly
settled enough to warrant being above v1.0.0 (that was mostly done to
denote the project was done in the context of the class it was
originally for).

Despite this being "Bad," I'm going to go back to pre-release
versioning (starting with v0.2.0), while I figure things out.